feat#9.5 add logic check item when delete category

+++ code/backend/src/modules/categories/categories.service.ts
@@ -1,8 +1,9 @@
-import { Injectable, Inject, InjectionToken, ConflictException, NotFoundException } from '@nestjs/common';
+import { Injectable, Inject, InjectionToken, ConflictException, NotFoundException, BadRequestException } from '@nestjs/common';
 import { Category } from './entities/category.entity.js';
 import type { ICategoryRepository } from './repositories/category.repository.interface.js';
 import { CreateCategoryDto } from './dto/create-category.dto.js';
 import { UpdateCategoryDto } from './dto/update-category.dto.js';
+import { IItemRepository } from '@modules/items/repositories/item.repo.interface.js';
 
 export const CATEGORY_REPOSITORY_TOKEN: InjectionToken<ICategoryRepository> 
   = Symbol('ICategoryRepository');
@@ -12,6 +13,7 @@ export class CategoriesService {
   constructor(
     @Inject(CATEGORY_REPOSITORY_TOKEN)
     private readonly categoryRepository: ICategoryRepository,
+    private readonly itemRepository: IItemRepository,
   ) {}
 
   async create(dto: CreateCategoryDto): Promise<Category> {
@@ -50,9 +52,9 @@ export class CategoriesService {
     if (!category) {
       throw new NotFoundException('Category not found');
     }
-    // TODO: Implement a check to see if there are any items associated with this category before deletion. If there are, throw an exception or handle it accordingly.
-    if(true){
-      console.warn('ItemService hasnt been implemented yet, skipping check for associated items');
+    const items = await this.itemRepository.findWithOptions({ where: { category: { id } } });
+    if (items.length > 0) {
+      throw new BadRequestException('Cannot delete category because items are linked to it');
     }
     await this.categoryRepository.delete(id);
   }
